+++ localization/enUS.lua
@@ -30,7 +30,7 @@ L["options.Random"] = "Random (instead of one by one)"
 L["options.Look"] = "Look at me! (Show off in resting areas)"
 L["options.Parachute"] = "Parachute mode - dismount anyway, no matter how high you fly and deep the fall"
 L["options.Together"] = "In group: Stay on ground together and don't summon flying mounts (except taxi)"
-L["options.PreferSwimming"] = "In water: Use water mounts, even if flying is possible"
+L["options.AlternateSwimming"] = "Whilst swimming alternate between water mounts and flying mounts"
 L["options.Taxi"] = "Taxi mode (Always summon taxi mount when in group)"
 L["options.Debug"] = "Debug mode"
 L["options.Autoopen"] = "Auto open and close with mount journal"
@@ -84,7 +84,7 @@ L["button.Help"] = "Help"
 
 L["Mount journal - Open Mounty"] = "Mounty #TLV"
 
-L["help"] = "Insert coin and try again!\n
+L["help"] = [[Insert coin and try again!\n
 /mounty  - to open the options frame\n
 /mounty magic - to let Mounty #TLV's magic happen\n
 /mounty ground - to summon ground mount\n
@@ -104,10 +104,14 @@ L["help"] = "Insert coin and try again!\n
 /mounty debug on|off (Show debug info)\n
 /mounty profile - to show the current used profile\n
 /mounty profile Heart - to switch to or create the profile Heart\n
-/mounty version - to show the current version of the addon"
+/mounty version - to show the current version of the addon]]
 
 L["quick.title"] = "Quick start"
-L["quick.text"] = "- Pull your favourite mounts from the mount journal into the categories of Mounty #TLV.\n- Open WoW's settings for key bindings, select Mounty #TLV and configure your magic key to mount.\n- Press your magic key to pick up the perfect mount for here and now.\n- Enjoy Mounty #TLV and take a look and all the other options and possibilities.\n\nTake a look at the README for all details."
+L["quick.text"] = [[- Pull your favourite mounts from the mount journal into the categories of Mounty #TLV.\n
+- Open WoW's settings for key bindings, select Mounty #TLV and configure your magic key to mount.\n
+- Press your magic key to pick up the perfect mount for here and now.\n
+- Enjoy Mounty #TLV and take a look and all the other options and possibilities.\n\n
+Take a look at the README for all details.]]
 
 L["readme.URL"] = "https://github.com/last-voice/mounty/blob/main/README.md"